add never option for locking the vault
