+++ mobile-app/src/screens/Auth/Register.tsx
@@ -18,7 +18,8 @@ import uuid from 'react-native-uuid'
 import { getDeviceName, getUniqueId } from 'react-native-device-info'
 import { generateEncryptionKey } from '../../../shared/generateEncryptionKey'
 import { DeviceContext } from '../../providers/DeviceProvider'
-import { IBackgroundStateSerializable, DeviceState } from '@utils/Device'
+import { IBackgroundStateSerializable } from '@utils/Device'
+import { DeviceState } from '@utils/DeviceState'
 import { saveAccessToken } from '@utils/tokenFromAsyncStorage'
 import SInfo from 'react-native-sensitive-info'
 import { Platform } from 'react-native'
