+++ mobile-app/src/utils/Device.ts
@@ -1,7 +1,6 @@
 import { apolloClient } from '../apollo/ApolloClient'
 import cryptoJS from 'crypto-js'
 import SInfo from 'react-native-sensitive-info'
-import { generateEncryptionKey } from '../../shared/generateEncryptionKey'
 import {
   EncryptedSecretGql,
   EncryptedSecretType,
@@ -9,17 +8,6 @@ import {
 } from '@shared/generated/graphqlBaseTypes'
 import { z, ZodError } from 'zod'
 import { loginCredentialsSchema } from './loginCredentialsSchema'
-import {
-  AddEncryptedSecretsDocument,
-  AddEncryptedSecretsMutation,
-  AddEncryptedSecretsMutationVariables,
-  MarkAsSyncedDocument,
-  MarkAsSyncedMutation,
-  MarkAsSyncedMutationVariables,
-  SyncEncryptedSecretsDocument,
-  SyncEncryptedSecretsQuery,
-  SyncEncryptedSecretsQueryVariables
-} from '@shared/graphql/ExtensionDevice.codegen'
 import messaging from '@react-native-firebase/messaging'
 import {
   LogoutDocument,
@@ -29,6 +17,7 @@ import {
 import { clearAccessToken } from './tokenFromAsyncStorage'
 import mitt from 'mitt'
 import { getDeviceName, getUniqueId } from 'react-native-device-info'
+import { DeviceState } from './DeviceState'
 
 export type SecretSerializedType = Pick<
   EncryptedSecretGql,
@@ -94,10 +83,11 @@ function getRandomInt(min: number, max: number) {
   return Math.floor(Math.random() * (max - min) + min) //The maximum is exclusive and the minimum is inclusive
 }
 type SecretTypeUnion = ILoginSecret | ITOTPSecret
-const isLoginSecret = (secret: SecretTypeUnion): secret is ILoginSecret =>
-  'loginCredentials' in secret
+export const isLoginSecret = (
+  secret: SecretTypeUnion
+): secret is ILoginSecret => 'loginCredentials' in secret
 
-const isTotpSecret = (secret: SecretTypeUnion): secret is ITOTPSecret =>
+export const isTotpSecret = (secret: SecretTypeUnion): secret is ITOTPSecret =>
   'totp' in secret
 
 export type AddSecretInput = Array<
@@ -106,254 +96,6 @@ export type AddSecretInput = Array<
     loginCredentials?: any
   }
 >
-
-// TODO move out of this file
-export class DeviceState implements IBackgroundStateSerializable {
-  decryptedSecrets: (ILoginSecret | ITOTPSecret)[]
-  constructor(parameters: IBackgroundStateSerializable) {
-    Object.assign(this, parameters)
-    this.decryptedSecrets = this.getAllSecretsDecrypted()
-  }
-  email: string
-  userId: string
-  deviceName: string
-
-  encryptionSalt: string
-  masterEncryptionKey: string
-  secrets: Array<SecretSerializedType>
-
-  authSecret: string
-  authSecretEncrypted!: string
-
-  //Settings
-  lockTime: number
-  syncTOTP: boolean
-  autofill: boolean
-  language: string
-  theme: string
-  biometricsEnabled = false
-
-  //Timer
-  lockTimeStart: number
-  lockTimeEnd: number
-  lockTimerRunning = false
-
-  setMasterEncryptionKey(masterPassword: string) {
-    this.masterEncryptionKey = generateEncryptionKey(
-      masterPassword,
-      this.encryptionSalt
-    )
-    this.save()
-  }
-
-  encrypt(stringToEncrypt: string) {
-    return cryptoJS.AES.encrypt(stringToEncrypt, this.masterEncryptionKey, {
-      iv: cryptoJS.enc.Utf8.parse(this.userId)
-    }).toString()
-  }
-  decrypt(encrypted: string) {
-    return cryptoJS.AES.decrypt(encrypted, this.masterEncryptionKey, {
-      iv: cryptoJS.enc.Utf8.parse(this.userId)
-    }).toString(cryptoJS.enc.Utf8)
-  }
-
-  async save() {
-    device.lockedState = null
-    this.decryptedSecrets = this.getAllSecretsDecrypted()
-
-    await SInfo.setItem(
-      'deviceState',
-      JSON.stringify({ backgroundState: this, lockedState: null }),
-      {
-        sharedPreferencesName: 'mySharedPrefs',
-        keychainService: 'myKeychain'
-      }
-    )
-  }
-
-  getSecretDecryptedById(id: string) {
-    const secret = this.decryptedSecrets.find((secret) => secret.id === id)
-    if (secret) {
-      return this.decryptSecret(secret)
-    }
-  }
-
-  getSecretsDecryptedByHostname(host: string) {
-    const secrets = this.decryptedSecrets.filter(
-      (secret) => host === new URL(secret.url ?? '').hostname
-    ) as Array<ILoginSecret | ITOTPSecret>
-    if (secrets) {
-      return secrets.map((secret) => {
-        return this.decryptSecret(secret)
-      })
-    }
-    return []
-  }
-
-  getAllSecretsDecrypted() {
-    return this.secrets.map((secret) => {
-      return this.decryptSecret(secret)
-    })
-  }
-
-  private decryptSecret(secret: SecretSerializedType) {
-    const decrypted = this.decrypt(secret.encrypted)
-    let secretDecrypted: ILoginSecret | ITOTPSecret
-    if (secret.kind === EncryptedSecretType.TOTP) {
-      secretDecrypted = {
-        ...secret,
-        totp: decrypted
-      } as ITOTPSecret
-    } else if (secret.kind === EncryptedSecretType.LOGIN_CREDENTIALS) {
-      const parsed = JSON.parse(decrypted)
-
-      try {
-        loginCredentialsSchema.parse(parsed.loginCredentials)
-        secretDecrypted = {
-          ...parsed,
-          ...secret
-        } as ILoginSecret
-      } catch (err: unknown) {
-        secretDecrypted = {
-          ...secret,
-          label: parsed.label,
-          url: parsed.url,
-          loginCredentials: {
-            username: '',
-            password: '',
-            parseError: err as Error
-          }
-        } as ILoginSecret
-      }
-    } else {
-      throw new Error('Unknown secret type')
-    }
-
-    return secretDecrypted
-  }
-
-  /**
-   * fetches newly added/deleted/updated secrets from the backend and updates the device state
-   */
-  async backendSync() {
-    const { data } = await apolloClient.query<
-      SyncEncryptedSecretsQuery,
-      SyncEncryptedSecretsQueryVariables
-    >({
-      query: SyncEncryptedSecretsDocument,
-      fetchPolicy: 'no-cache'
-    })
-
-    if (data) {
-      const deviceState = device.state
-      if (data && deviceState) {
-        const backendRemovedSecrets =
-          data.currentDevice.encryptedSecretsToSync.filter(
-            ({ deletedAt }) => deletedAt
-          )
-        const newAndUpdatedSecrets =
-          data.currentDevice.encryptedSecretsToSync.filter(
-            ({ deletedAt }) => !deletedAt
-          )
-
-        const secretsBeforeSync = deviceState.secrets
-        const unchangedSecrets = secretsBeforeSync.filter(
-          ({ id }) =>
-            !backendRemovedSecrets.find(
-              (removedSecret) => id === removedSecret.id
-            ) &&
-            !newAndUpdatedSecrets.find(
-              (updatedSecret) => id === updatedSecret.id
-            )
-        )
-
-        deviceState.secrets = [...unchangedSecrets, ...newAndUpdatedSecrets]
-
-        await this.save()
-
-        await apolloClient.mutate<
-          MarkAsSyncedMutation,
-          MarkAsSyncedMutationVariables
-        >({ mutation: MarkAsSyncedDocument })
-
-        const actuallyRemovedOnThisDevice = backendRemovedSecrets.filter(
-          ({ id: removedId }) => {
-            return secretsBeforeSync.find(({ id }) => removedId === id)
-          }
-        )
-        console.log(
-          '~ actuallyRemovedOnThisDevice',
-          actuallyRemovedOnThisDevice
-        )
-        return {
-          removedSecrets: actuallyRemovedOnThisDevice.length,
-          newAndUpdatedSecrets: newAndUpdatedSecrets.length
-        }
-      }
-    }
-  }
-
-  findExistingSecret(secret) {
-    const existingSecretsOnHostname = this.getSecretsDecryptedByHostname(
-      new URL(secret.url).hostname
-    )
-
-    return existingSecretsOnHostname.find(
-      (s) =>
-        (isLoginSecret(s) &&
-          s.loginCredentials.username === secret.loginCredentials?.username) ||
-        (isTotpSecret(s) && s.totp === secret.totp)
-    )
-  }
-
-  /**
-   * invokes the backend mutation and pushes the new secret to the bgState
-   * @param secret
-   * @returns the added secret
-   */
-  async addSecrets(secrets: AddSecretInput) {
-    // const existingSecret = this.findExistingSecret(secrets)
-    // if (existingSecret) {
-    //   return null
-    // }
-
-    const { data } = await apolloClient.mutate<
-      AddEncryptedSecretsMutation,
-      AddEncryptedSecretsMutationVariables
-    >({
-      mutation: AddEncryptedSecretsDocument,
-      variables: {
-        secrets: secrets.map((secret) => {
-          const stringToEncrypt =
-            secret.kind === EncryptedSecretType.TOTP
-              ? secret.totp
-              : JSON.stringify(secret.loginCredentials)
-
-          const encrypted = this.encrypt(stringToEncrypt as string)
-
-          return {
-            encrypted,
-            kind: secret.kind
-          }
-        })
-      }
-    })
-    if (!data) {
-      throw new Error('failed to save secret')
-    }
-    console.log('saved secret to the backend', secrets)
-    const secretsAdded = data.me.addEncryptedSecrets
-
-    this.secrets.push(...secretsAdded)
-    await this.save()
-    return secretsAdded
-  }
-
-  async removeSecret(secretId: string) {
-    this.secrets = this.secrets.filter((s) => s.id !== secretId)
-    this.save()
-  }
-}
 
 export class Device {
   state: DeviceState | null = null
@@ -425,23 +167,26 @@ export class Device {
   }
 
   syncSettings(config: SettingsInput) {
-    this.state!.autofill = config.autofill
-    this.state!.lockTime = config.vaultLockTimeoutSeconds
-    this.state!.syncTOTP = config.syncTOTP
-    this.state!.language = config.language
-    this.state!.theme = config.theme
+    const { state } = this
+    if (!state) {
+      throw new Error('cannot sync without state')
+    }
+    state.autofill = config.autofill
+    state.lockTime = config.vaultLockTimeoutSeconds
+    state.syncTOTP = config.syncTOTP
+    state.language = config.language
+    state.theme = config.theme
 
     // Sync timer
-    if (this.state!.lockTime !== config.vaultLockTimeoutSeconds) {
-      this.state!.lockTimeEnd =
-        Date.now() + config.vaultLockTimeoutSeconds * 1000
+    if (state.lockTime !== config.vaultLockTimeoutSeconds) {
+      state.lockTimeEnd = Date.now() + config.vaultLockTimeoutSeconds * 1000
     }
 
-    if (device.state!.lockTimeEnd <= Date.now()) {
+    if (state.lockTimeEnd <= Date.now()) {
       device.lock()
-    } else if (device.state?.lockTimeEnd) {
-      if (!this.lockInterval) {
-        console.log('lockTimeEnd', device.state?.lockTimeEnd)
+    } else if (state?.lockTimeEnd) {
+      if (!this.lockInterval && state.lockTime > 0) {
+        console.log('lockTimeEnd', state?.lockTimeEnd)
         device.startVaultLockTimer()
       }
     }
@@ -489,7 +234,7 @@ export class Device {
   }
 
   async lock() {
-    this.clearInterval()
+    this.clearLockInterval()
 
     if (!this.state) {
       return
@@ -552,7 +297,7 @@ export class Device {
   }
 
   clearAndReload = async () => {
-    this.clearInterval()
+    this.clearLockInterval()
     await clearAccessToken()
     await device.clearLocalStorage()
     this.emitter.emit('stateChange')
@@ -614,7 +359,19 @@ export class Device {
     }, 5000)
   }
 
-  clearInterval = () => {
+  setLockTime(lockTime: number) {
+    this.state!.lockTime = lockTime
+
+    this.clearLockInterval()
+    if (lockTime > 0) {
+      this.state!.lockTimeEnd = Date.now() + this.state!.lockTime * 1000
+      this.startVaultLockTimer()
+    }
+
+    this.save(false)
+  }
+
+  clearLockInterval = () => {
     this.lockInterval = clearInterval(this.lockInterval!)
   }
 }
